Move code into package jsonTransforms.
