+++ src/components/jsonTransforms/usfmjsStructureRules.js
